+++ src/main/java/by/siarhiejbahdaniec/playeradsplugin/config/ConfigHolder.java
@@ -10,7 +10,7 @@ public interface ConfigHolder {
     @NotNull
     public String getString(String key, String def);
 
-    public long getLong(String key);
+    public int getInt(String key);
 
     public boolean getBoolean(String key);
 }
